+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # CHANGELOG for jbox
 
 ----------------------------------------------------------------------
+## v1.1.0 (snapshot)
+
+- Support stdin and stdout streaming.
+- Add integration tests in the `itest` directory.
+
+
+----------------------------------------------------------------------
 ## v1.0.0 (1 May 2018)
 
-Initial release that supports encrypt and decrypt.
-
+- Initial release that supports encrypt and decrypt.
 
 
 ----------------------------------------------------------------------